+++ loki-core/src/main/java/io/github/guoshiqiufeng/loki/core/handler/impl/RocketMqRemotingHandler.java
@@ -16,28 +16,19 @@
 package io.github.guoshiqiufeng.loki.core.handler.impl;
 
 import io.github.guoshiqiufeng.loki.MessageContent;
-import io.github.guoshiqiufeng.loki.core.config.ConsumerConfig;
 import io.github.guoshiqiufeng.loki.core.handler.AbstractHandler;
 import io.github.guoshiqiufeng.loki.core.handler.HandlerHolder;
-import io.github.guoshiqiufeng.loki.core.toolkit.ThreadPoolUtils;
 import io.github.guoshiqiufeng.loki.enums.MqType;
 import io.github.guoshiqiufeng.loki.support.core.config.LokiProperties;
-import io.github.guoshiqiufeng.loki.support.core.consumer.ConsumerRecord;
-import io.github.guoshiqiufeng.loki.support.core.pipeline.PipelineUtils;
+import io.github.guoshiqiufeng.loki.support.core.consumer.ConsumerConfig;
 import io.github.guoshiqiufeng.loki.support.core.producer.ProducerRecord;
 import io.github.guoshiqiufeng.loki.support.core.producer.ProducerResult;
 import io.github.guoshiqiufeng.loki.support.core.util.StringUtils;
 import io.github.guoshiqiufeng.loki.support.rocketmq.remoting.RocketRemotingClient;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.rocketmq.client.consumer.DefaultMQPushConsumer;
-import org.apache.rocketmq.client.consumer.listener.ConsumeConcurrentlyStatus;
-import org.apache.rocketmq.client.consumer.listener.MessageListenerConcurrently;
-import org.apache.rocketmq.common.message.MessageExt;
-import org.apache.rocketmq.shaded.com.google.common.base.Throwables;
 
 import java.util.Arrays;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutorService;
 import java.util.function.Function;
 
 /**
@@ -145,82 +136,7 @@ public class RocketMqRemotingHandler extends AbstractHandler {
      */
     @Override
     public void pushMessageListener(ConsumerConfig consumerConfig, Function<MessageContent<String>, Void> function) {
-        String topic = consumerConfig.getTopic();
-        String topicPattern = consumerConfig.getTopicPattern();
-        String tag = consumerConfig.getTag();
-        if (StringUtils.isEmpty(topic) && StringUtils.isEmpty(topicPattern)) {
-            if (log.isErrorEnabled()) {
-                log.error("RocketMqHandler# pushMessageListener error: topic and topicPattern is both null");
-            }
-            return;
-        }
-        try {
-            if (StringUtils.isEmpty(tag)) {
-                tag = "*";
-            }
-            ExecutorService executorService = ThreadPoolUtils.getSingleThreadPool();
-            String finalTag = tag;
-            DefaultMQPushConsumer consumer = rocketRemotingClient.getConsumer(consumerConfig.getConsumerGroup(), consumerConfig.getIndex());
-            CompletableFuture.runAsync(() -> {
-                if (!StringUtils.isEmpty(topicPattern)) {
-
-                } else {
-                    try {
-                        consumer.subscribe(topic, finalTag);
-                        consumer.registerMessageListener((MessageListenerConcurrently) (msgs, context) -> {
-                            try {
-                                for (MessageExt msgExt : msgs) {
-                                    if (log.isDebugEnabled()) {
-                                        log.debug("msgExt:{}", msgExt);
-                                    }
-                                    ConsumerRecord consumerRecord = covertConsumerRecord(msgExt);
-                                    consumerRecord = PipelineUtils.processListener(consumerRecord);
-                                    if (consumerRecord == null) {
-                                        return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
-                                    }
-                                    function.apply(new MessageContent<String>()
-                                            .setMessageId(consumerRecord.getMessageId())
-                                            //.setMessageGroup(msgExt.getProperty(MessageConst.PROPERTY_PRODUCER_GROUP))
-                                            .setTopic(consumerRecord.getTopic())
-                                            .setTag(consumerRecord.getTag())
-                                            .setKeys(consumerRecord.getKeys())
-                                            .setBody(consumerRecord.getBodyMessage())
-                                            .setBodyMessage(consumerRecord.getBodyMessage()));
-                                }
-                            } catch (Exception e) {
-                                if (log.isErrorEnabled()) {
-                                    log.error("RocketMqHandler# pushMessageListener error:{}", Throwables.getStackTraceAsString(e));
-                                }
-                                return ConsumeConcurrentlyStatus.RECONSUME_LATER;
-                            }
-                            return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
-                        });
-                        consumer.start();
-                    } catch (Exception e) {
-                        if (log.isErrorEnabled()) {
-                            log.error("Exception consumer in registerMessageListener: {}", e.getMessage());
-                        }
-                    }
-                }
-
-            }, executorService).exceptionally(throwable -> {
-                if (log.isErrorEnabled()) {
-                    log.error("Exception occurred in CompletableFuture: {}", throwable.getMessage());
-                }
-                return null;
-            });
-
-        } catch (Exception e) {
-            if (log.isErrorEnabled()) {
-                log.error("RocketMqHandler# pushMessageListener error:{}", e.getMessage());
-            }
-            throw new RuntimeException(e);
-        }
+        rocketRemotingClient.consumer(consumerConfig, function);
     }
 
-    private ConsumerRecord covertConsumerRecord(MessageExt msgExt) {
-        return new ConsumerRecord(msgExt.getTopic(), msgExt.getTags(),
-                msgExt.getMsgId(), null, Arrays.asList(msgExt.getKeys().split(",")),
-                new String(msgExt.getBody()));
-    }
 }